Extract mapping between Stripe errors and domain errors to an internal extension. Nobody needs to know
